Revert "docstrings for io and misc"
This reverts commit f4c70b46a0466455ae2833168b3acc9f282dbe0e.
